correct

+++ python/dz2/task1.py
@@ -1,4 +1,4 @@
 print('Input side of the rectangle:')
-a, b = map(int, input().split())
+a, b = map(float, input().split())
 print('The area of the rectangle is', a * b)
 print('The rectangle perimeter is', (a + b) * 2)
